Refactor `highlight_selection`
- Move the Selection stuff to its own module
- Rename highlight_selection to highlight_keys
